[Minor] Moved routes around, update product mrp modal
